+++ spec/integration/build_spec.rb
@@ -59,26 +59,30 @@ RSpec.describe '`metal build`' do
   end
 
   def run_command(command, &block)
-      Timeout::timeout 5 do
-        Open3.popen3 command do |stdin, stdout, stderr, thread|
+    Timeout::timeout 5 do
+      Open3.popen3 command do |stdin, stdout, stderr, thread|
+        begin
+          pid = thread.pid
+          block.call(stdin, stdout, stderr, pid)
+        rescue Exception => e
           begin
-            pid = thread.pid
-            block.call(stdin, stdout, stderr, pid)
-          rescue Exception => e
-            begin
-              # Try to read output `stdout` and `stderr`, or just ensure original
-              # exception raised if not available.
-              max_bytes_to_read = 30000
-              stdout_data = stdout.read_nonblock(max_bytes_to_read)
-              stderr_data = stderr.read_nonblock(max_bytes_to_read)
-              puts "stdout:\n#{stdout_data}\n\nstderr:\n#{stderr_data}"
-            rescue
-              raise e
-            end
-            raise
+            stdout_data = read_io_stream(stdout)
+            stderr_data = read_io_stream(stderr)
+            puts "stdout:\n#{stdout_data}\n\nstderr:\n#{stderr_data}"
+          rescue
+            raise e
           end
+          raise
         end
+      end
     end
+  end
+
+  def read_io_stream(stream)
+    max_bytes_to_read = 30000
+    stream.read_nonblock(max_bytes_to_read)
+  rescue EOFError
+    ''
   end
 
   def expect_clears_up_built_node_marker_files