+++ src/config.rb
@@ -33,6 +33,7 @@ module Metalware
     # XXX DRY these paths up.
     KEYS_WITH_DEFAULTS = {
       build_poll_sleep: 10,
+      answer_files_path: '/var/lib/metalware/answers',
       built_nodes_storage_path: '/var/lib/metalware/cache/built-nodes',
       rendered_files_path: '/var/lib/metalware/rendered',
       pxelinux_cfg_path: '/var/lib/tftpboot/pxelinux.cfg',